Moved api out to server Reworked sockets to use socketio Added progress to nodes Added highlight to active node Added preview to saveimage node
